Create initial local storage values
I have set the default value of the note title and note body to display
when the page is loaded.  Next, I will create a function that saves what
the user types using localStorage.

I also moved the check and trash icons to the jumbotron for now.  Was
having an issue where it would overwrite those with the default value of
localStorage.  I will address this in a future update.